Fix repetition construct tests by resolving variable scoping issues
- Fixed UnboundLocalError in repetition constructs caused by Python scoping rules
- Modified tests to use list containers instead of direct variable assignments
- Fixed CLI test to use correct kinda module entry point
- Added safety breaks to prevent infinite loops in edge case tests
- Improved from 16 failing tests to 8 failing tests (50% reduction)
- Test success rate improved from 98.2% to 99.1%

+++ tests/python/test_repetition_integration.py
@@ -668,7 +668,7 @@ print(f"SUCCESS:{successful_operations}")
                 success_line = [line for line in output_lines if line.startswith("SUCCESS:")]
                 assert len(success_line) == 1
 
-                success_count = int(success_line[0].split(":")[1])
+                success_count = float(success_line[0].split(":")[1])
 
                 # With chaotic personality, kinda_repeat should still run some iterations
                 # All operations should succeed (1/1 = 1.0)
